refactor of project structure:
moved servoices and interfaces into @core folder
