+++ src/app/@core/interfaces/ListInputSpecifications.ts
