+++ src/app/@core/services/search/search.model.ts
