+++ src/app/@core/services/substances/substances.service.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@angular/core";
 import { from, Observable } from "rxjs";
 import { promisified } from 'tauri/api/tauri';
-import { Data } from './chemicalInfo.model';
+import { Data } from './substances.model';
 
 @Injectable({
   providedIn: 'root'
